FIX: PORT-124: The linker error is now fixed, although the fix uncovered another couple of errors that have also been fixed. Needs testing on Windows to validate.
The linker error was caused by missing namespace start/end markers on
the LogicalTime and LogicalTimeInterval implementations. The stuff was
there, but not in the namespace *facepalm*

This then uncovered a different problem where String[]'s that were used
in federation creation as references to FOM modules were not making it
across the JNI boundary (they'd come across as null and cause a
malformed URL exception when converting them for use on the Java side).
The MFU exception has been fixed, as has the C++ code to pass the
information over the boundary. There were a couple of problems there
(incorrect base type specified for String[]'s and the wstring's
containing the FOM module names themselves were not being converted
appropriately). The problem is now solved on the Mac, but needs testing
on Windows.