PORT-128: Added in defaults for the sent and received ordering values, as well as the transport handles. These try to get as close as I can while still guessing based on the situation (for example, the sent and received ordering is RELIABLE if the message is delivered TSO) but I have not put in a full implementation. The defaults have been added to all overloads of the following methods:
 * reflectAttributeValues()
 * receiveInteraction()
 * removeObjectInstance()

It's too much of a stretch to calculate the proper values inside the
callback handler (performance hit as the RTI has to reach back into its
stores to do more processing). When done properly, these should be done
by the generic handlers further back in the chain, but there is a bit of
work to do there so I'm just choosing sensisble defaults for now given
that the primary target audience of Portico is the training domain where
these values are ignored anyway (in fact, I'm trying to think of ANY
domain where these values are used).