Fix bug that truncates large attribute updates at 989 bytes.
The `ObjectInput.read()` call being used to fetch attribute data from
the incoming buffer was only reading the first 989 bytes. We had
pre-allocated a buffer with the correct size, so this resulted in the
appearance of truncation after this position. Changing the call to
`readFully()` ensured that the full amount of data available was being
read.

Change has been made for both the UpdateAttributes and SendInteraction
message classes. Unit tests for sizes of 1MB confirm the behaviour.

Fixes: #65