Allow the WAN Server to collect metrics and log them as CSV
Added support to the `Host` class for logging metrics on the way through
execution, with the full data being written to a .cvs file whenever the
host exits.

Data currently captured at an internal of messages sent. Each 100
messages, the number of messages sent is captures, as is the number of
messages that is currently in the queue.

Working on: #135