Enable bundling on WAN connections for high throughput performance
This commit adds proper bundling support to the LocalGateway so that
messages sent over the WAN can be grouped togethe to make more efficient
use of network resources. Testing locally has shown that with bundling
enabled, throughput tests are able to ready ~80% of the local multicast
performance with two federates.

The previously seen memory exhaustion issues have also started to be
controlled. Bundling means fewer messages are sent, and although they
are larger, the overall send time vice individual messages is much
smaller. As such, fewer messages end up queued and the router appears to
reach a steady state. To provide some extra room, the WAN Router startup
scripts increase the heap memory slightly, although this did not seem to
be fully necessary and is a safety precaution only.

Working on: #135