Change the way handlers are loaded to remove dynamic classpath lookup.
Previously, handlers were loaded by searching certain parts of the classpath
for the @MessageHandler annotation. The "certain parts" bit is what caused
most of the problem. We couldn't search the full classpath, because in many
cases it is enormous. As such, we tried to limit the search locations to
the location of portico.jar (to get the main plugins that make up the RTI)
and a dedicated "plugins" folder.

This is all fine as long as you can find portico.jar. The failure to find
the jar is often perfectly valid (expanded code on classpath, Eclipse builds
etc...). As such, not finding it isn't an error and if the environment is
properly configured, the core plugins will be found via a dedicated plugin
folder. This is already sounding complicated. To make it worse, when lookup
failed you wouldn't get an error (because it might be valid), instead you'd
get a cryptic "null returned for sink" or some such error caused by there
being no handler to deal with a given message because it was never found.

Anyway, long story short - nobody was using the plugins mechanism, and
frankly, if anyone did they'd have other issues so I am just taking it away.
In its place I've put in a MessageRegistry class that just does the work of
manually registering the core handlers. Simple, in fact it is down right
"not clever" in most ways... which means it works.

fin.

Fix: #157