Add partial MOM implementation for 1516e.
Name resolution now works, and updates are generated in response to
`RequestAttributeUpdate` and `RequestClassAttributeUpdate` requests.

Basic MOM support added for 1516e. The MOM hierarchy is now manually
added to a federation so that handles can be controlled, and names are
resolved properly. Federates can request updates and they will receive
the responses, but currently the formatting of the attribute reflections
is not correct (just follows the "everything as a string" approach).

Updates to the tests and implementation needed to ensure that the
encoding is correct.

Working: #55