Add initial support for MOM to 1516e
Many of the values are not currently supported, but the basics are now
complete for 1516e MOM support, including the proper serialization of
those attribtutes of `HLAfederate` and `HLAfederation` that are
supported.

New test additions have been included, as have refactorings of the major
MOM data providers to deliver values in format that is properly encoded
depending on the version of the interface being used. This has also
triggered some changes through the 1516-2000 tests to match this.

Fix: #55