Fixing (i hope) problem with hanging test
testConnectWhenAlreadyConnected() appears to be hanging intermittently
on Travis builds. This never happens for me locally and getting logs out
of that environment can be hard, especially as the problem is
intermittent (maybe 1/10 builds). Shuffling some things around to
hopefully cause the condition this test is confirming to execute
earlier, before anything has the chance to hang. If the tests starts
failing on local builds a better investigation is needed.