Fix problem with HLAprivilegeToDelete not getting added to FOM.
Due to recent change in MOM processing, the MIM is no longer loaded from
disk, but rather built in code. This lets us give specific handles,
which allows us to look up attributes quickly without using their name
(and in turn, allows us to do things in a HLA v1.3 & 1516e compliant
way).

This commit fixes the missing attribute problem by adding it on
validation if it is missing.

Working: #175