Added HLAvariantRecord encoding and decoding
Added IEEE 1516e HLA Variant Record encoding and decoding. Unit tests
(inspired by the tests for HLA1516eFixedRecord) are included and pass.

Required implementation of hashCode and equals on HLA1516eOctet. It
will be necessary to implement hashCode and equals on any type that
will be used as a variant record discriminant.