Implementation of HLA1516eVariantRecord (#178)
* Added HLAvariantRecord encoding and decoding

Added IEEE 1516e HLA Variant Record encoding and decoding. Unit tests
(inspired by the tests for HLA1516eFixedRecord) are included and pass.

Required implementation of hashCode and equals on HLA1516eOctet. It
will be necessary to implement hashCode and equals on any type that
will be used as a variant record discriminant.

* Added and removed some commentary

* Commentary and small code changes

Mainly a commentary commit. Identified logic errors in getOctetBoundary
(discriminant octet boundary is not considered anymore) and decode
(wasn’t handling the case where there was no variant for the decoded
discriminant).

* Deleted errant import auto inserted by my IDE.