Prevent throwing of FederationDoesNotExist for JGroups and fix problems with associated tests that is caused by it.
FederationExecutionDoesNotExist exceptions were being thrown because
with JGroups, a federation ceases to exist when all federates resign. So
if the last one out is a good citizen and tried to clean up, it will get
a false exception. This in turn was causing probelms with some tests. In
addition to this, there was an automatic disconnection on destroy issue
that was preventing tests from passing with the JGroups binding. This
has been commented out now, so explict disconnection is still required.

Fix: #189 - mostly. It doesn't really "fix" the problem, it just removes
the negative impact of the way Portico works. Given the scope of the
remaining problem (it isn't "technically correct" - that's about the
only consequence) this is good enough.