Ensure fake resign messages are queued for crashed federates.
This process previously required that a federate already have been
"suspected" of having crashed. Now it works regardless.

Every time the view is updated we store the set of those joined members.
If there are members who were in the previous view but are no longer in
this one, we process it as a resignation of that federate.

Working: #201