Send Goodbye messages from coordinator for any crashed federates.
The previous fix to deal with the propagation of crash information
won't work over a WAN as it relies on listening to changes to the local
JGroups view (which doesn't propagate).

This update forces the coordinator who detects the crash to send a
Goodbye message (which will propagate) to all joined members.

Working: #201