Remove the use of ServiceRegistry in the standard/header 1516e classes.
The use of ServiceRegistry in Java 9 is now hard limited to only a
particular set of Java IO/Image types. The 1516e standard headers used
it for a generic lookup mechanism, which frankly was a bit of a smell
(take one look at the package the classes were in and you know they were
not meant for this purpose).

This change removes that code and hard-codes the return of
Portico-specific implementations. Fine because it'll all be hidden to
the user if they change RTIs.

Fix: #204