Remove the "Dtx" RTIamb extension for 1516e and put in a general version
Building on the old "JLC" apporach from the past, I have added an
"extensions" package to the standard 1516e Java interfaces. This
contains a single RTIambassadorEx interface that extends the default
RTIambassador and adds our custom methods.

This change is mirrored inside the implementation code and I have
flicked the default Rti1516eAmbassador to abstract to ensure that we
only ever pass the extended version. So now:

  hla.rti1516e.RTIambassador >> org.portico.impl.hla1516e.Rti1516eAmbassador (abstract)
  hla.rti1516e.extensions.RTIambassadorEx >>  org.port....Rti1516eAmbassadorEx (concrete)

Working: #209