Fix an NPE in the JGroups binding when joining a federation you created.
When we create a federation we create our own manifest instance rather
than receiving it from the coordinator (as there is none prior to our
join). However, we never actually add the ObjectModel instance to the
local Manifest when we do this, hence an NPE when trying to access it
later on.

*facepalm*