Update the CPP ieee1516e example to match the java
I have updated the ieee1516e examples to match more closely. There are
still a couple of things that are different. I added an additional
interaction
class to the fed to be able to keep in an example of passing a parameter
to an interaction class. The example could be refined more but it is
enough for me to be able to add in attributes test.