(MAJOR REFACTOR) Move from fully distributed to a central RTI process.
This is part of major ongoing work to support #219. We are moving to a
central RTI process where the administrative tasks are all completed
inside the RTI, but high-volume messages like data exchanges are still
routed over efficient, multicast networks.

We are also adding in more robust network-bridging unicast support that
will replace the WAN Forwarder and support filtering of messages based
on rules configurable inside the RID.

This commit is the first major step. It adds a temporary portico2 folder
that replicates a lot of existing structure and brings into place a
separate RTI, LRC and Forwarder component that serve each of the major
subsystems. It currently only supports 1516e and still have many broken
parts, however the infrastructure and core services are present and it
is robust enough to support further development by more than one person.

This commit represents a number of previous WIP changes squashed into
one bulk of work. Due to the above, the individual changes are extremely
broad and too numerious to list individually. Rather, this consolidated
commit just marks the state from which a more managed process should be
undertaken.

Working: #219