Add interface to enable attribute/parameter datatype queries
This collection of interfaces enables the user to query the datatype
of parameter and attributes. The interfaces include Basic, Simple,
Enumerated, Array, FixedRecord, VariantRecord and NA datatypes.
The main class that does the grunt work is the DatatypeRetrieval
class.

Variation from the Java Interface

Accessor functionality for the Datatype interfaces.
To keep data security and meet polymorphism constraints for the
cpp compilation I have removed the setters. this data should
not be modified by the end user.

Removed IEnumerator interface.
The additional layer of abstraction was not required for the cpp
portion... I think.

Variation from the Java Additions

Beside additions to the Java Proxy class, there was an additional class added
for encoding the FOM datatypes into an XML object. Currently, this only
exports the datatypes but should eventually have the rest of the FOM added.

Dependencies

For the FOM xml parsing, the pugixml library has been used.
The .lib files and include have been added to the build process.

Unit Tests

The most basic unit test framework and DatatypeRetrieval tests have been
added to the VS solution and build. It includes the testbed that is used
for the HLA1.3 spec except updated for 1516e