Add interface to enable attribute/parameter datatype queries (#262)
* Add interface to enable attribute/parameter datatype queries

This collection of interfaces enables the user to query the datatype
of parameter and attributes. The interfaces include Basic, Simple,
Enumerated, Array, FixedRecord, VariantRecord and NA datatypes.
The main class that does the grunt work is the DatatypeRetrieval
class.

Variation from the Java Interface

Accessor functionality for the Datatype interfaces.
To keep data security and meet polymorphism constraints for the
cpp compilation I have removed the setters. this data should
not be modified by the end user.

Dependencies

For the FOM xml parsing, the pugixml library has been used.
The .lib files and include have been added to the build process.

Unit Tests

The most basic unit test framework and DatatypeRetrieval tests have been
added to the VS solution and build. It includes the testbed that is used
for the HLA1.3 spec except updated for 1516e

* PORT-239 Edits from PR review

- Adds ant compile target for the ieee1516e unit test suite
- Adds ant test target for the ieee1516e unit test suite
- Reverts standard headers RTI/1516.h and RTI/RTIambassadorFactory.h
  back to their original state
- Adds RTI_EXPORT qualifiers to all new public interfaces
- Moved RTIambassadorEx out of standard ieee1516e namespace into
  portico1516e namespace
- Moved all new datatype related interfaces into portico1516e namespace
- Made the datatype unit tests a bit more thorough. They now work on the
  RPR fom and do a deep analysis of the various properties of the
  datatype
- Removes constructor from RTIambassadorEx interface
- DatatypeRetrieval has been reduced down to one public method
  getDatatype() and will self-initialize
- DatatypeRetrieval's xpath query for datatype nodes was incorrectly
  capturing <enumerator> nodes as they also have a name attribute.
  Modified the datatype node lookup to use several queries, specific
  to the path of each datatype class
- Edited inline documentation where appropriate
- Corrected indentation and formatting where appropriate
- Corrected memory management annotations (user is NOT responsible for
  memory management of datatype related pointers, they are cached
  internally be the LRC)
- Removed getFom() from the RTIambassadorEx interface
- Guarded against delete of uninitialised listenerFederate pointer in
  TestQuickCreate

Other important things that are not directly related to the PR

- Renamed JniUtils::toShortString() to JniUtils::toNarrowString()
- ieee1516e Runtime::generateWinPath() will now fallback to JAVA_HOME if
  no JRE can be found at RTI_HOME
- ieee1516e Runtime::getHlaVersion() now correctly sets JVM argument
  portico.cpp.hlaversion to ieee1516e