Add initial commit of encryption infrastructure.
Not much actual encryption going on at the moment despite a number of
tests having been coded and removed. Now moving onto the stage where I
need to put the configuration in place and this will require a bit of a
refactor of the configuration files themselves (moving to XML, sigh).

Working: #240, #244