Overhaul the structure of the header to be more efficient and streamlined.
This is a partial commit of WIP. It overhauls the structure of the
header to enable more efficient processing and a better separation of
both Encryption/Authentication, and to add a new message call type
called a "Notification". This replaces the Async control messages of the
past, which were also confusing as they implied a certain value of
"isFromRti" in the Portico message itself. Now we have:

  - Data Messages: As they were
  - Control Request: Between single RTI and Federate
  - Control Response: Response to a Control Request
  - Notification: Unsolicited message sent from the RTI

Further updates to come will split Authentication and Encryption
(currently the PKI and Symmetric packages do not do this well) and will
move the RIT to an XML format.