Introduce the WIP base structure of an XML configuration for the RID
This moves the RID configuration from a properties file to an XML one.
With the recent changes there has been the introduction of a log more
nested configuration information. This doesn't translate well to the
properties format and it can be confusing. As such, I'm introducing an
XML configuration file. The configuration code is a little more...
effort, but the actual configuration file itself is pretty simple.

This commit has the base infrastructure for most of it and cuts the RTI
RID parser over to use the XML version. Tasks to complete:

  - Move LRC and Forwarder objects over to this format
  - Remove the old properties based configuration classes
  - Move the current XML files out of the .xml sub-package
  - Add support for command-line args overrides
  - Add support for RTIambassador::connect() overrides