Add privToDelete to ObjectRoot if it is missing (1516e)
The ObjectModel class has a method called setObjectRoot(OCMetadata).
This method now checks to make sure that HLAprivilegeToDelete is present
as an attribute of the class and adds it if it is not.

We could get exceptions when the root module we are using for whatever
reason does not have it (as is the case in our 1516e example modules)