Tie the period of time a JGroups federate will wait for coordinator discovery to the response timeout configuration property
Previously we were waiting for a hard coded period of time. Now this
value is tied to the value of the JGroups configuration RID property
portico.jgroups.responseTimeout. This allows it to be increased in WAN
environments where waits might be significantly longer.

Will need to look at potentially short-circuiting the wait when we
receive a response early for coordinator discovery. Object registration
with a name will also be as slow as whatever this value is configured to
be.