Fix bug that caused parts of 1516e MIM to be missing from federations
Previously we were stripping any parts of the MOM that came in from
federates and then replacing it with the bits of the MOM that we
supported, only with specific, controllable handles. The problem is that
we don't support the full MOM, and while at it, stripping parts of the
object model that the user has provided is probably a bit rude.

This commit fixes that. The MIM is now used as the base for EVERY
federation and the parts of the RTI that service MOM requests have been
updated to deal with the fact that the handles are no longer static and
must be looked up for each federation.