JGroups federates now always take an updated manifest from coordinator
Previously, when a JGroups federate joined a federation, they'd tell the
whole federate and each individual federate would track the changes. But
this can lead to inconsistencies in handles assigned to parts of the FOM
unless all modules are merged in exactly the same order.

This update changes things so that all federates ONLY take model updates
from the federation coordinator to ensure consistency.