Trigger resign for messages that fail RoleCall
Previously, if a federate failed the role call, it would still show up
in other federate's list of active federates, and the MOM object for it
would still be floating around. This change means that if a federate
fails to join the federation due to a role call error (which for JGroups
is AFTER the initial join handshake), it will then send a resign that
should cause the other federates to clean out any record of it.