Improves cross-version FOM name lookup and ensure MOM loading for HLA1.3
Updates the cross-version name lookup to ignore _any_ instance of "HLA"
as the prefix for a class name. E.g. allows lookup of `Manager.Federate`
to find `HLAmanager.HLAfederate`.

Updates federation creation handling so that if the MOM is missing from
an HLA v1.3 fed file on load, the 1516e MIM is loaded and used in its
place. This, combined with the cross-version name lookup improvements
will mean the 1516e standard MIM will be available in v1.3 federates if
the FED file they include does not have MOM provisions.