Fix linux c++ sandbox target dependencies
If I specify the fully qualified target name for the C++ compile targets
in the sandbox target, they get remapped to the wrong target when
resolving bad dependency links created by Ant's include/import task
prefixing all target names and their dependencies.

Our patched version of Ant is meant to fix this, but there is a bug.

When <include> is invoked, "cpp.hla13.compile" gets changed to
"cpp.hla13.cpp.hla13.compile" (standard Ant behaviour). Our patched
version of Ant then remaps any missing dependencies, but rather than
finding the target inside the file (cpp.hla13.compile) it finds the
_root_ "compile" target, which isn't what we want.

This is a bug in CppTask that needs to be fixed, but for now we can
work around it by not using the qualified name for targets inside the
same build file. Doing this means the targets still resolve even when
the Ant import/include task runs and so no remapping is needed, and the
bug isn't triggered.