Remove custom C++ extensions for datatype introspection
These were added to allow introspection on the aggregate FOM built up
from potentially several unknown federates. Rather than being _in_ the
RTI, this should be a separate, standalone library where the surface
area of interface change to the RTI is extremely limited to just extract
the FOM structure as a string (or something like that) so that the
portico RTI interface itself is kept light.