Reduce message bundling default timeout from 30ms to 5ms
Message bundling is enabled by default, and the default timeout is 30ms.
This value is too high for all but the most chatty federations, and has
the impact of increasing default latency. Reducing it down to 5ms feels
like it balances responsiveness with a desire to not generate a packet
flood on networks.

Fix: #349