Update example scripts so javac calls enable annotation processing (#352)
This isn't strictly necessary, but without it <javac> will issue
warnings. We have lots of libraries on the classpath that use annotation
processing, so this warning will forever pop up. The least we can do is
show how to get rid of the warning. Without annotation processing the
RTI as it exists today may well break (won't be able to discover
handlers), so it can't hurt to just enable it explicitly.