Fix bug preventing JGroups federates from updating FOMs after peer join
The FOM is set in the LRCState when a federate joins a federation. If
another federate subsequently joins the federation and extends the FOM,
and it happens to add modules, peer JGroups federate would never have
their local FOM updated, so they never knew the new types existed.

This update fixes that and pushes in new object models received as a
result of remote peer joins.

Working: #188